requirements.txt
numpy~=1.23.4
scipy~=1.10.1
pyvista~=0.44.2
backports.strenum~=1.3.1; python_version < '3.11'